+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/bindings/wod/WodBindingDeprecationProblem.java
@@ -1,15 +1,13 @@
 package org.objectstyle.wolips.bindings.wod;
 
 import org.eclipse.jface.text.Position;
-import org.objectstyle.wolips.bindings.api.IApiBinding;
 
 /**
- * Class representing a deprecated binding.
- * 
- * @author jw
+ * A WOD validation problem for bindings whose values resolve to deprecated
+ * Java members.
  */
 public class WodBindingDeprecationProblem extends WodBindingProblem {
-  public WodBindingDeprecationProblem(IWodElement element, IApiBinding binding, String bindingName, String message, Position position, int lineNumber, boolean warning) {
-    super(element, binding, bindingName, message, position, lineNumber, warning);
+  public WodBindingDeprecationProblem(IWodElement element, String bindingName, String message, Position position, int lineNumber, boolean warning) {
+    super(element, bindingName, message, position, lineNumber, warning);
   }
 }