+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/componenteditor/inspector/BindingsPopUpMenu.java
@@ -13,9 +13,9 @@ import org.eclipse.swt.graphics.Point;
 import org.eclipse.swt.widgets.Decorations;
 import org.eclipse.swt.widgets.Menu;
 import org.eclipse.swt.widgets.MenuItem;
-import org.objectstyle.wolips.bindings.api.IApiBinding;
 import org.objectstyle.wolips.bindings.api.ApiSnapshot;
 import org.objectstyle.wolips.bindings.wod.IWodElement;
+import org.objectstyle.wolips.bindings.wod.VisibleBinding;
 import org.objectstyle.wolips.bindings.wod.WodProblem;
 import org.objectstyle.wolips.componenteditor.ComponenteditorPlugin;
 import org.objectstyle.wolips.wodclipse.core.completion.WodParserCache;
@@ -57,11 +57,11 @@ public class BindingsPopUpMenu implements MenuListener {
 		boolean showMenu = false;
 		ApiSnapshot api = wodElement.getApi(_cache.getJavaProject(), WodParserCache.getTypeCache());
 		if (api != null) {
-			IApiBinding[] apiBindings = wodElement.getApiBindings(api);
-			if (apiBindings != null && apiBindings.length > 0) {
-				Set<IApiBinding> keyBindings = new TreeSet<IApiBinding>();
-				Set<IApiBinding> actionBindings = new TreeSet<IApiBinding>();
-				for (IApiBinding binding : apiBindings) {
+			VisibleBinding[] visibleBindings = wodElement.getVisibleBindings(api);
+			if (visibleBindings != null && visibleBindings.length > 0) {
+				Set<VisibleBinding> keyBindings = new TreeSet<VisibleBinding>();
+				Set<VisibleBinding> actionBindings = new TreeSet<VisibleBinding>();
+				for (VisibleBinding binding : visibleBindings) {
 					if (binding.isAction()) {
 						actionBindings.add(binding);
 						showMenu = true;
@@ -78,14 +78,14 @@ public class BindingsPopUpMenu implements MenuListener {
 				List<WodProblem> wodProblems = WodModelUtils.getProblems(wodElement, _cache);
 
 				BindingSelectionListener selectionListener = new BindingSelectionListener(wodElement, droppedKeyPath, _cache);
-				for (IApiBinding keyBinding : keyBindings) {
+				for (VisibleBinding keyBinding : keyBindings) {
 					MenuItem mi = createMenuItem(wodElement, keyBinding, wodProblems);
 					mi.addSelectionListener(selectionListener);
 				}
 				if (!keyBindings.isEmpty() && !actionBindings.isEmpty()) {
 					new MenuItem(_menu, SWT.SEPARATOR);
 				}
-				for (IApiBinding actionBinding : actionBindings) {
+				for (VisibleBinding actionBinding : actionBindings) {
 					MenuItem mi = createMenuItem(wodElement, actionBinding, wodProblems);
 					mi.addSelectionListener(selectionListener);
 				}
@@ -103,13 +103,13 @@ public class BindingsPopUpMenu implements MenuListener {
 		return showMenu;
 	}
 
-	protected MenuItem createMenuItem(IWodElement element, IApiBinding binding, List<WodProblem> wodProblems) {
+	protected MenuItem createMenuItem(IWodElement element, VisibleBinding binding, List<WodProblem> wodProblems) {
 		MenuItem menuItem = new MenuItem(_menu, SWT.NONE);
 		menuItem.setData(binding);
 		menuItem.setText(binding.getName());
 		if (element.getBindingNamed(binding.getName()) != null) {
 			menuItem.setImage(ComponenteditorPlugin.getDefault().getImage(ComponenteditorPlugin.CONNECTED_ICON));
-		} else if (WodModelUtils.hasValidationProblem(binding, wodProblems)) {
+		} else if (WodModelUtils.hasValidationProblem(binding.getName(), wodProblems)) {
 			menuItem.setImage(ComponenteditorPlugin.getDefault().getImage(ComponenteditorPlugin.UNCONNECTED_PROBLEM_ICON));
 		} else {
 			menuItem.setImage(ComponenteditorPlugin.getDefault().getImage(ComponenteditorPlugin.UNCONNECTED_ICON));
@@ -137,10 +137,10 @@ public class BindingsPopUpMenu implements MenuListener {
 		public void widgetSelected(SelectionEvent event) {
 			MenuItem item = (MenuItem) event.widget;
 			item.setImage(ComponenteditorPlugin.getDefault().getImage(ComponenteditorPlugin.CONNECTED_ICON));
-			IApiBinding apiBinding = (IApiBinding) item.getData();
+			VisibleBinding vb = (VisibleBinding) item.getData();
 			RefactoringWodElement refactoringWodElement = new RefactoringWodElement(_wodElement, _cache);
 			try {
-				refactoringWodElement.setValueForBinding(_droppedKeyPath, apiBinding.getName());
+				refactoringWodElement.setValueForBinding(_droppedKeyPath, vb.getName());
 			} catch (Exception e) {
 				e.printStackTrace();
 				ComponenteditorPlugin.getDefault().log("Failed to add binding.", e);