+++ wolips/plugins/ng.componenteditor/java/org/objectstyle/wolips/wodclipse/core/document/WodBindingValueHyperlink.java
@@ -7,29 +7,30 @@ import org.eclipse.jface.text.IRegion;
 import org.eclipse.jface.text.Position;
 import org.eclipse.jface.text.Region;
 import org.eclipse.jface.text.hyperlink.IHyperlink;
+import org.objectstyle.wolips.bindings.api.ApiUtils;
 import org.objectstyle.wolips.bindings.wod.IWodBinding;
 import org.objectstyle.wolips.bindings.wod.IWodElement;
-import org.objectstyle.wolips.bindings.wod.TypeCache;
 import org.objectstyle.wolips.locate.LocateException;
 import org.objectstyle.wolips.wodclipse.core.Activator;
 import org.objectstyle.wolips.wodclipse.core.completion.WodCompletionUtils;
 import org.objectstyle.wolips.wodclipse.core.completion.WodParserCache;
 
+/**
+ * Hyperlink for binding values in WOD/template editors. Ctrl+click on a
+ * binding value opens the corresponding Java source, or offers to create
+ * a key/action if the keypath doesn't exist.
+ */
 public class WodBindingValueHyperlink implements IHyperlink {
-  private IWodElement _element;
-  private IWodBinding _binding;
-  private TypeCache _cache;
   private IRegion _region;
   private String _bindingValue;
+  private boolean _isAction;
   private IType _componentType;
 
-  public WodBindingValueHyperlink(IWodElement element, IWodBinding binding, IRegion region, String bindingValue, IType componentType, TypeCache cache) {
-    _element = element;
-    _binding = binding;
+  public WodBindingValueHyperlink(IRegion region, String bindingValue, boolean isAction, IType componentType) {
     _region = region;
     _bindingValue = bindingValue;
+    _isAction = isAction;
     _componentType = componentType;
-    _cache = cache;
   }
 
   public IRegion getHyperlinkRegion() {
@@ -46,13 +47,17 @@ public class WodBindingValueHyperlink implements IHyperlink {
 
   public void open() {
     try {
-      WodCompletionUtils.openBinding(_bindingValue, _binding, _componentType, false);
+      WodCompletionUtils.openBinding(_bindingValue, _isAction, _componentType, false);
     }
     catch (Exception ex) {
       Activator.getDefault().log(ex);
     }
   }
 
+  /**
+   * Creates a hyperlink for the given binding's value, or returns null if
+   * the binding doesn't exist or isn't a keypath.
+   */
   public static WodBindingValueHyperlink toBindingValueHyperlink(IWodElement wodElement, String bindingName, WodParserCache cache) throws JavaModelException, CoreException, LocateException {
     WodBindingValueHyperlink hyperlink = null;
     IWodBinding wodBinding = wodElement.getBindingNamed(bindingName);
@@ -62,7 +67,8 @@ public class WodBindingValueHyperlink implements IHyperlink {
         Region elementRegion = new Region(valuePosition.getOffset(), valuePosition.getLength());
         IType componentType = cache.getComponentType();
         if (componentType != null) {
-          hyperlink = new WodBindingValueHyperlink(wodElement, wodBinding, elementRegion, wodBinding.getValue(), componentType, WodParserCache.getTypeCache());
+          boolean isAction = ApiUtils.isActionBindingName(wodBinding.getName());
+          hyperlink = new WodBindingValueHyperlink(elementRegion, wodBinding.getValue(), isAction, componentType);
         }
       }
     }